% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{baby_count}
\alias{baby_count}
\title{baby_count S3 object}
\usage{
baby_count(name, start_year, vec)
}
\arguments{
\item{name}{The name of the baby}

\item{start_year}{The year at which the vector of counts starts}

\item{vec}{The vector of counts by year}
}
\value{
The baby_count object
}
\description{
baby_count S3 object
}
\examples{
df <- data.frame(Betty=c(73, 62, 86), Murial=c(93, 63, 27))
rownames(df) <- c(1974:1976)
baby_count(colnames(df)[1], as.numeric(rownames(df)[1]), df[[1]])
}
