% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_change_functions.R
\name{calculate_percent_change}
\alias{calculate_percent_change}
\title{Calculate Percent Changes}
\usage{
calculate_percent_change(counts)
}
\arguments{
\item{counts}{A vector of sequential counts}
}
\value{
A vector of percent changes, with the first being 0
}
\description{
Calculate Percent Changes
}
\examples{
calculate_percent_change(c(1:10))
}
