% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_media_dates.R
\name{correlate_poi_and_media}
\alias{correlate_poi_and_media}
\title{Correlate media release year with top percent change baby name yearsmerge pc_df and media df by name??}
\usage{
correlate_poi_and_media(pc_df)
}
\description{
@description
This function take in the pc_df data frame and the media_df with release_year
column added and creates a final data frame with the poi_years for our media
events.

@param pc_df Percent change data frame with names, year, percent_change
columns.
@param media_df Filtered media data frame that has release_date
@returns Data frame with new column that corresponds to years that are
points of interest, poi_year. poi_years are defined as the top 20
greatest percent change from year to year. 1 for poi_year = TRUE 0
otherwise.final_df has columns: name, title, release_year, poi_year,
and percent_change.
}
