% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_media_dates.R
\name{create_media_names_df}
\alias{create_media_names_df}
\title{Create Media Names Dataframe}
\usage{
create_media_names_df(poi_df, media_df, max_range = 5)
}
\arguments{
\item{poi_df}{A data frame that contains columns name, year, percent change.}

\item{media_df}{A data frame with columns name, title, and year.}

\item{max_range}{The maximum range between a release year and a poi year to
consider significant}
}
\value{
The media_influenced_names data frame with name, title, release_year,
poi_year, and percent_change columns.
}
\description{
Correlates names from poi_df and characters from media_df when the year from
poi_df is within 0 and max_range years after the year from media_df
}
