% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_change_functions.R
\name{create_percent_change_df}
\alias{create_percent_change_df}
\title{Create Percent Change Dataframe}
\usage{
create_percent_change_df(count_df)
}
\arguments{
\item{count_df}{A dataframe with columns "name" and a bunch of years, where
the years columns contain a count of names within that year.}
}
\value{
A dataframe with columns "name" and a bunch of years, where the years
columns contain a percent change from the last year.
}
\description{
Creates a dataframe from a names dataframe with percent change year over year
rather than yearly counts
}
