% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_change_functions.R
\name{create_poi_df}
\alias{create_poi_df}
\title{Create Point of Interest Dataframe}
\usage{
create_poi_df(pc_df, pc_cutoff = 100)
}
\arguments{
\item{pc_df}{The percent change dataframe from which to find points of
interest}

\item{pc_cutoff}{The cutoff above which percent changes (positive or
negative) are interesting}
}
\value{
A dataframe consisting of three columns: name, year, and
percent_change.
}
\description{
Create Point of Interest Dataframe
}
