% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter_baby_names_by_char}
\alias{filter_baby_names_by_char}
\title{Function that takes in the movie and baby names data frames and filters and
returns the filtered baby names data frame.}
\usage{
filter_baby_names_by_char(baby_names, movie_titles)
}
\arguments{
\item{baby_names}{The data frame with baby names information.}

\item{movie_titles}{The data frame with movie character information}
}
\value{
The filtered baby names data frame with only names
that have an associated character in the movie titles data frame.
}
\description{
Function that takes in the movie and baby names data frames and filters and
returns the filtered baby names data frame.
}
