% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter_baby_names_by_names}
\alias{filter_baby_names_by_names}
\title{Filter Baby Names By Names}
\usage{
filter_baby_names_by_names(filter_df, filter_by_df, multiple = 1)
}
\arguments{
\item{filter_df}{The dataframe to filter}

\item{filter_by_df}{The dataframe to filter by}

\item{multiple}{The ratio of commonness. Values over 1 will allow names to
be included even if they are more common in the other dataframe, while
values uner 1 will eliminate names even if they are less common in the
other data frame.}
}
\value{
The filtered dataframe
}
\description{
Filters a dataframe of names by another dataframe and elminates all names
that are more common in the other dataframe.
}
