% Generated by roxygen2: do not edit by hand

% Please edit documentation in R/filter_funcs_char_baby.R
\name{filter_char_by_baby_names}
\alias{filter_char_by_baby_names}
\title{Function that takes in the movie and baby names data frames and filters and
returns the filtered movie data frame.}
\usage{
filter_char_by_baby_names(movie_titles, baby_names)
}
\arguments{
\item{movie_titles}{The data frame with movie character information}

\item{baby_names}{The data frame with baby names information.}
}
\value{
The filtered movie character data frame with only movie characters
that have a name listed in the baby names data frame.
}
\description{
Function that takes in the movie and baby names data frames and filters and
returns the filtered movie data frame.
}
