% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{influences}
\alias{influences}
\title{influences S3 class}
\usage{
influences(df)
}
\arguments{
\item{df}{The dataframe to convert to an influences object. Should have five
columns: name, title, release_year, poi_year, and percent_change.}
}
\value{
The influences object
}
\description{
influences S3 class
}
\examples{
df <- data.frame(name=c("Betty","Betty"), title=c("Jumanji", "Phantom"),
  release_year=c(1996, 1990), poi_year=c(1996, 1991),
  percent_change=c(154.6532, -233.9874))
influences(df)

}
