% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{male_babies}
\alias{male_babies}
\title{Male Baby Name Counts}
\format{
\subsection{\code{MALE_BABIES}}{

A data frame with 50 rows and 39,079 columns:
\describe{
\item{name}{There is an individual column for each unique male baby name}
\item{year}{The rownames are the years of 1973 to 2022 and contain the
counts of the baby names for that year}
}
}
}
\source{
\url{https://www.ssa.gov/oact/babynames/limits.html}
}
\usage{
male_babies
}
\description{
A consolidated dataset from the Social Security Administration website
consisting of male baby names and counts from 1973 to 2022.
}
\keyword{datasets}
