% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_media_dates.R
\name{media_influenced_names}
\alias{media_influenced_names}
\title{Media Influenced Names Function}
\usage{
media_influenced_names(poi_df, media_df)
}
\arguments{
\item{poi_df}{A data frame that contains the year, percent change,
baby names from the create_poi_df function.}

\item{media_df}{A data frame with the name, title, and release year for the
media data.}
}
\value{
The media_influenced_names data frame with name, title, release_year,
poi_year, and percent_change columns.
}
\description{
A function that creates a media_influenced_names data frame.
}
