% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_influences.R
\name{plot.influences}
\alias{plot.influences}
\title{Generic Plot Class influences}
\usage{
\method{plot}{influences}(x, y, sex, ...)
}
\arguments{
\item{x}{An object of class influences.}

\item{y}{An object of class baby_count or baby_pc.}

\item{sex}{Enter whether the name is male or female.}

\item{...}{Traditional plot function arguments.}
}
\value{
A plot of the influences object
}
\description{
This function uses the generic plot function to plot an object of class
influences a long with one additional class object (baby_count or baby_pc).
}
\examples{
# creating baby_pc object
df <- data.frame(Leia=c(50, -4.8, 5, 366.7, 59.2, -24.4, 37.3),
Kizzy=c(0, 0, 0, 111600, -59.1, -41, -55.8))
rownames(df) <- c(1974:1980)
pc_obj <- baby_pc(colnames(df)[1], as.numeric(rownames(df)[1]), df[[1]])
# creating baby_count object
df <- data.frame(Leia=c(21, 20, 21, 98, 156, 118, 162),
Kizzy=c(0, 0, 0, 1116, 456, 269, 119))
rownames(df) <- c(1974:1980)
bc_obj <- baby_count(colnames(df)[1], as.numeric(rownames(df)[1]), df[[1]])
# creating influences object
df <- data.frame(name=c("Leia", "Leia"), title=c("Star Wars", "Star Wars"),
release_year=c(1977, 1978), poi_year=c(1977, 1978),
percent_change=c(366.7, 59.2))
influ_obj <- influences(df)
# plotting objects
plot(influ_obj, pc_obj)
plot(influ_obj, bc_obj)
}
