% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{validate_baby_count}
\alias{validate_baby_count}
\title{Validator for baby_count constructor}
\usage{
validate_baby_count(name, start_year, vec)
}
\arguments{
\item{name}{The name of the baby}

\item{start_year}{The year at which the vector of counts starts}

\item{vec}{The vector of baby counts per year}
}
\value{
The validated vector
}
\description{
Validator for baby_count constructor
}
\examples{
df <- data.frame(Betty=c(73, 62, 86), Murial=c(93, 63, 27))
rownames(df) <- c(1974:1976)
validate_baby_count(colnames(df)[1], as.numeric(rownames(df)[1]), df[[1]])
}
