% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{validate_baby_pc}
\alias{validate_baby_pc}
\title{Validator for baby_pc constructor}
\usage{
validate_baby_pc(name, start_year, vec)
}
\arguments{
\item{name}{The name of the baby}

\item{start_year}{The year at which the vector of percent changes starts}

\item{vec}{The vector of percent changes per year}
}
\value{
The validated vector
}
\description{
Validator for baby_pc constructor
}
\examples{
df <- data.frame(Betty=c(-73.5, -62.7, 86.1), Murial=c(93, 63, -27))
rownames(df) <- c(1974:1976)
validate_baby_pc(colnames(df)[1], rownames(df)[1], df[[1]])
}
