% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{validate_influences}
\alias{validate_influences}
\title{Validator for influences class}
\usage{
validate_influences(df)
}
\arguments{
\item{df}{The dataframe to validate. Should have five
columns: name, title, release_year, poi_year, and percent_change.}
}
\value{
The validated dataframe
}
\description{
Validator for influences class
}
