% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{female_babies}
\alias{female_babies}
\title{Female Baby Name Counts}
\format{
\subsection{\code{FEMALE_BABIES}}{

A data frame with 50 rows and 61,622 columns:
\describe{
\item{name}{There is an individual column for each unique female baby name}
\item{year}{The rownames are the years of 1973 to 2022 and contain the
counts of the baby names for that year}
}
}
}
\source{
\url{https://www.ssa.gov/oact/babynames/limits.html}
}
\usage{
female_babies
}
\description{
A consolidated dataset from the Social Security Administration website
consisting of female baby names and counts from 1973 to 2022.
}
\keyword{datasets}
