% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{male_babies}
\alias{male_babies}
\title{Male Baby Names}
\format{
\subsection{\code{male_babies}}{

A data frame with 50 rows and 61622 columns:
\describe{
\item{names}{The counts of male babies given a particular name, with each
row being a year}
}
}
}
\source{
\url{https://developer.imdb.com/non-commercial-datasets/}
}
\usage{
male_babies
}
\description{
A subset of data from the Social Security Administration consisting of names
given to female babies born 1973 - 2023.
}
\keyword{datasets}
